LSurvival

Survival analysis functions in Julia for time-to-event outcomes that can include:

Loss-to-follow-up/right censoring

Late entry/left truncation (not available in Survival.jl)

"Person-period" data structures (not available in Survival.jl)

Observation weights (not available in Survival.jl)

Competing risks (not available in Survival.jl)

Capabilities include estimators for

Kaplan-Meier non-parametric conditional risk functions

Aalen-Johansen non-parametric cause-specific unconditional risk functions

Cox proportional hazards model (Efron's or Breslow's methods for ties)

Convenience functions enable:

Non-parametric bootstrapping, cluster-bootstrapping

Estimating baseline hazards from a Cox Model

Estimating cause-specific risk from an exhaustive set of Cox models for competing risk outcomes

Simple simulation of competing and non-competing survival events

Cluster robust variance estimation (without bootstrapping)

Martingale, score, Schoenfeld, and dfbeta residuals

Cluster robust variance estimation

Plans to include:

Parametric survival models

Stratification in Cox models

The package has been tuned to follow the "survival" package from R in terms of specific estimators/results.

Report issues here

Installation

using Pkg; Pkg.add(url = "https://github.com/alexpkeil1/LSurvival.jl")

Quick examples

Single event type: Cox model and Kaplan-Meier curve

using Random, LSurvival, Distributions, LinearAlgebra

# generate some data under a discrete hazards model
expit(mu) = inv(1.0 + exp(-mu))

function dgm(rng, n, maxT; regimefun = int_0)
    V = rand(rng, n)
    LAY = Array{Float64,2}(undef, n * maxT, 4)
    keep = ones(Bool, n * maxT)
    id = sort(reduce(vcat, fill(collect(1:n), maxT)))
    time = (reduce(vcat, fill(collect(1:maxT), n)))
    for i = 1:n
        v = V[i]
        l = 0
        a = 0
        lkeep = true
        for t = 1:maxT
            currIDX = (i - 1) * maxT + t
            l = expit(-3 + 2 * v + 0 * l + 0 * a) > rand(rng) ? 1 : 0
            a = 0.1 > rand(rng) ? 1 : 0
            y = expit(-3 + 2 * v + 0 * l + 2 * a) > rand(rng) ? 1 : 0
            LAY[currIDX, :] .= [v, l, a, y]
            keep[currIDX] = lkeep
            lkeep = (!lkeep || (y == 1)) ? false : true
        end
    end
    id[findall(keep)], time[findall(keep)] .- 1, time[findall(keep)], LAY[findall(keep), :]
end

id, int, outt, data = dgm(MersenneTwister(), 1000, 10; regimefun = int_0)
data[:, 1] = round.(data[:, 1], digits = 3)
d, X = data[:, 4], data[:, 1:3]
wt = rand(length(d)) # random weights just to demonstrate usage

# Cox model
# Breslow's partial likelihood
m = fit(PHModel, X, int, outt, d, ties = "breslow", wts = wt)

# Efron's partial likelihood
m2 = fit(PHModel, X, int, outt, d, ties = "efron", wts = wt)

#equivalent way to specify 
# using `coxph` function
m2b = coxph(X, int, outt, d, ties = "efron", wts = wt)

# using `coxph` function with `Tables.jl` and `StatsAPI.@formula` interface (similar to GLM.jl)
tab = ( in = int, out = out, d=d, x=X[:,1], z1=X[:,2], z2=X[:,3]) # can also be a DataFrame from DataFrames.jl
m2b = coxph(@formula(Surv(in, out, d)~x+z1+z2), ties = "efron", wts = wt)

# can also be done if there is no late entry
m2b = coxph(@formula(Surv(out, d)~x+z1+z2), ties = "efron", wts = wt)
# can also be done if there is no late entry and no right censoring (i.e. all times are failure times)
m2b = coxph(@formula(Surv(out)~x+z1+z2), ties = "efron", wts = wt)

# Kaplan-Meier estimator of the cumulative risk/survival
res = kaplan_meier(int, outt, d)

Competing event analysis: Aalen-Johansen and Cox-model-based estimators of the cumulative risk/survival

using Random, LSurvival, Distributions, LinearAlgebra

# simulate some data
function dgm_comprisk(; n = 100, rng = MersenneTwister())
    z = rand(rng, n) .* 5
    x = rand(rng, n) .* 5
    dt1 = Weibull.(fill(0.75, n), inv.(exp.(-x .- z)))
    dt2 = Weibull.(fill(0.75, n), inv.(exp.(-x .- z)))
    t01 = rand.(rng, dt1)
    t02 = rand.(rng, dt2)
    t0 = min.(t01, t02)
    t = Array{Float64,1}(undef, n)
    for i = 1:n
        t[i] = t0[i] > 1.0 ? 1.0 : t0[i]
    end
    d = (t .== t0)
    event = (t .== t01) .+ 2.0 .* (t .== t02)
    wtu = rand(rng, n) .* 5.0
    wt = wtu ./ mean(wtu)
    reshape(round.(z, digits = 4), (n, 1)),
    reshape(round.(x, digits = 4), (n, 1)),
    round.(t, digits = 4),
    d,
    event,
    round.(wt, digits = 4)
end

z, x, t, d, event, wt = dgm_comprisk(; n = 100, rng = MersenneTwister(12))
X = hcat(x,z)
enter = t .* rand(length(d))*0.02 # create some fake entry times

# Aalen-Johansen estimator: marginal cause-specific risks
res_aj = aalen_johansen(enter, t, event; wts = wt);
res_aj

# Cox-model estimator: cause-specific risks at given levels of covariates
fit1 = fit(PHModel, X, enter, t, (event .== 1), ties = "efron",  wts = wt)
#n2idx = findall(event .!= 1)
n2idx = findall(event .> -1)
fit2 = fit(PHModel, X[n2idx,:], enter[n2idx], t[n2idx], (event[n2idx] .== 2), ties = "breslow",  wts = wt[n2idx])

# risk at referent levels of `x` and `z` (can be very extreme if referent levels are unlikely/unobservable)
res_cph_ref = risk_from_coxphmodels([fit1,fit2])

# risk at average levels of `x` and `z`
mnx = sum(x)/length(x)
mnz = sum(z)/length(z)
res_cph = risk_from_coxphmodels([fit1,fit2], coef_vectors=[coef(ft1), coef(ft2)], pred_profile=mean(X, dims=1))
# compare to Aalen-Johansen fit
res_aj

# this approach operates on left censored outcomes (which operate in the background in model fitting)
LSurvivalResp(enter, t, d, origintime=0)
LSurvivalCompResp(enter, t, event) # automatically infers origin

# can use the ID type to refer to units with multiple observations
id, int, outt, data = dgm(MersenneTwister(), 1000, 10; regimefun = int_0)
LSurvivalResp(int, outt, data[:,4], ID.(id))

Index of functions

LSurvival.AbstractLSurvivalParms

LSurvival.AbstractLSurvivalResp

LSurvival.AbstractNPSurv

LSurvival.AbstractPH

LSurvival.ID

LSurvival.LSurvivalCompResp

LSurvival.LSurvivalResp

LSurvival.PHModel

LSurvival.PHSurv

LSurvival.Strata

LSurvival._update_PHParms!

LSurvival.aalen_johansen

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.coxph

LSurvival.dgm

LSurvival.dgm_comprisk

LSurvival.jackknife

LSurvival.jackknife

LSurvival.jackknife

LSurvival.kaplan_meier

LSurvival.lgh_breslow!

LSurvival.lgh_efron!

LSurvival.risk_from_coxphmodels

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RecipesBase.apply_recipe

RecipesBase.apply_recipe

StatsAPI.confint

StatsAPI.confint

StatsAPI.confint

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit!

StatsAPI.loglikelihood

StatsAPI.nullloglikelihood

StatsAPI.residuals

StatsAPI.stderror

StatsAPI.stderror

StatsAPI.vcov

Function help

LSurvival.AbstractLSurvivalParms — Type

AbstractLsurvParms

Abstract type representing a model predictors and coefficient parameters

LSurvival.AbstractLSurvivalResp — Type

AbstractLsurvResp

Abstract type representing a model response vector

LSurvival.AbstractNPSurv — Type

Abstract type for non-parametric survival models, including Kaplan-Meier, Aalen Johansen, and Cox-model based estimates of survival using an Aalen-Johansen-like estimator

LSurvival.AbstractPH — Type

Abstract type for proportional hazards models

LSurvival.ID — Type

Type for identifying individuals in survival outcomes.

Used for the id argument in

Outcome types: LSurvivalResp, LSurvivalCompResp

Model types: PHModel, KMRisk, AJRisk

Accepts any Number or String. There is no significance to having this particular struct, but it enables easier use of multiple dispatch.

[ID(i) for i in 1:10]

LSurvival.LSurvivalCompResp — Type

Outcome type for competing risk survival outcomes subject to left truncation and right censoring (not generally needed for users)

Parameters

enter Time at observation start

exit Time at observation end

y event occurrence in observation

wts observation weights

eventtimes unique event times

origin origin on the time scale

id person level identifier (must be wrapped in ID() function)

eventtypes vector of unique event types

eventmatrix matrix of indicators on the observation level

Signatures:

julia struct LSurvivalCompResp{ E<:AbstractVector, X<:AbstractVector, Y<:AbstractVector, W<:AbstractVector, T<:Real, I<:AbstractLSurvivalID, V<:AbstractVector, M<:AbstractMatrix, } <: AbstractLSurvivalResp enter::E exit::X y::Y wts::W eventtimes::X origin::T id::Vector{I} eventtypes::V eventmatrix::M end

julia LSurvivalCompResp( enter::E, exit::X, y::Y, wts::W, id::Vector{I} )

julia LSurvivalCompResp( enter::E, exit::X, y::Y, id::Vector{I} )

julia LSurvivalCompResp( enter::E, exit::X, y::Y, wts::W, )

julia LSurvivalCompResp( enter::E, exit::X, y::Y, )

julia LSurvivalCompResp( exit::X, y::Y, ) where {X<:Vector,Y<:Union{Vector{<:Real},BitVector}}

LSurvival.LSurvivalResp — Type

Outcome type for survival outcome subject to left truncation and right censoring.

Will not generally be needed by users

Parameters

enter: Time at observation start

exit: Time at observation end

y: event occurrence in observation

wts: observation weights

eventtimes: unique event times

origin: origin on the time scale

id: person level identifier (must be wrapped in ID() function)

```julia struct LSurvivalResp{ E<:AbstractVector, X<:AbstractVector, Y<:AbstractVector, W<:AbstractVector, T<:Real, I<:AbstractLSurvivalID, } <: AbstractLSurvivalResp enter::E exit::X y::Y wts::W eventtimes::E origin::T id::Vector{I} end

```

julia LSurvivalResp( enter::E, exit::X, y::Y, wts::W, id::Vector{I}, ) where { E<:Vector, X<:Vector, Y<:Union{Vector{<:Real},BitVector}, W<:Vector, I<:AbstractLSurvivalID, }

```julia LSurvivalResp( enter::E, exit::X, y::Y, id::Vector{I}, )

```

julia LSurvivalResp( y::Vector{Y}, wts::W, id::Vector{I}, ) where {Y<:AbstractSurvTime,W<:Vector,I<:AbstractLSurvivalID}

julia LSurvivalResp( enter::E, exit::X, y::Y, ) where {E<:Vector,X<:Vector,Y<:Union{Vector{<:Real},BitVector}}

julia LSurvivalResp(exit::X, y::Y) where {X<:Vector,Y<:Vector}

Examples

```julia

no late entry

LSurvivalResp([.5, .6], [1,0])

```

LSurvival.PHModel — Type

PHModel: Mutable object type for proportional hazards regression (not generally needed for users)

Parameters

R Survival response

P # parameters

ties String: "efron" or "breslow"

fit Bool: logical for whether the model has been fitted

bh AbstractMatrix: baseline hazard estimates

Signatures

```julia mutable struct PHModel{G<:LSurvivalResp,L<:AbstractLSurvivalParms} <: AbstractPH R::G # Survival response P::L # parameters ties::String #"efron" or"breslow" fit::Bool bh::AbstractMatrix end

PHModel( R::G, P::L, ties::String, fit::Bool, ) where {G<:LSurvivalResp,L<:AbstractLSurvivalParms} PHModel(R::G, P::L, ties::String) where {G<:LSurvivalResp,L<:AbstractLSurvivalParms} PHModel(R::G, P::L) where {G<:LSurvivalResp,L<:AbstractLSurvivalParms} ``` Methods: fit, coef, confint, std_err, show

Example

```@example using LSurvival using Random import LSurvival: stepcox!, dgmcomprisk

z,x,t,d, event,wt = dgmcomprisk(MersenneTwister(1212), 100); enter = zeros(length(t)); X = hcat(x,z); R = LSurvivalResp(enter, t, Int.(d), wt) P = PHParms(X) mf = PHModel(R,P) LSurvival.fit!(mf) ```

LSurvival.PHSurv — Type

Mutable type for proportional hazards models (not generally needed by users)

PHSsurv: Object type for proportional hazards regression

surv::Vector{Float64} risk::Matrix{Float64} basehaz::Vector{Float64} event::Vector{Float64}

fitlist: vector of PHSurv objects (Cox model fits)

eventtypes: vector of unique event types

times: unique event times

surv: Overall survival at each time

risk: Cause-specific risk at each time (1 for each outcome type)

basehaz: baseline hazard for a specific event type

event: value of event type that occurred at each time

Methods: fit, show

mutable struct PHSurv{G<:Array{T} where {T<:PHModel}} <: AbstractNPSurv
fitlist::G        
eventtypes::AbstractVector
times::AbstractVector
surv::Vector{Float64}
risk::Matrix{Float64}
basehaz::Vector{Float64}
event::Vector{Float64}
end

PHSurv(fitlist::Array{T}, eventtypes) where {T<:PHModel}
PHSurv(fitlist::Array{T}) where {T<:PHModel}

LSurvival.Strata — Type

Type for identifying individuals in survival outcomes. Used for the strata argument in PHModel (not yet implemented)

Accepts any Number or String. There is no significance to having this particular struct, but it enables easier use of multiple dispatch.

[Strata(i) for i in 1:10]

LSurvival._update_PHParms! — Method

Update the partial likelihood, gradient and Hessian values from a Cox model fit (used during fitting, not generally useful for users).

Uses Breslow's or Efron's partial likelihood.

Updates over all observations

Signature

_update_PHParms!(
 m::M,
 # big indexes
 ne::I,
 caseidxs::Vector{Vector{T}},
 risksetidxs::Vector{Vector{T}},
 ) where {M<:AbstractPH,I<:Int,T<:Int}

updatePHParms!(m, risksetidxs, caseidxs, ne, den)

LSurvival.aalen_johansen — Method

Aalen-Johansen estimator for cumulative cause-specific risk (in the presence of competing events)

Signatures

 StatsBase.fit!(m::T; kwargs...) where {T<:AbstractNPSurv}

 aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; kwargs...)

Keyword arguments

wts::Vector{<:Real} = similar(enter, 0); vector of case weights (or zero length vector) for each observation

id::Vector{<:AbstractLSurvivalID} = [ID(i) for i in eachindex(y)]; Vector of AbstractSurvID objects denoting observations that form a single unit (used in bootstrap and jackknife methods)

atol = 0.00000001; absolute tolerance for defining tied event times

keepy = true; keep the outcome vector after fitting (may save memory with large datasets)

eps = 0.00000001; deprecated (replaced by atol)

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter, t, event, wts=wt)

LSurvival.bootstrap — Method

Bootstrapping coefficients of a proportional hazards model

Signatures

# single bootstrap draw, keeping the entire object
bootstrap(rng::MersenneTwister, m::PHModel)
bootstrap(m::PHModel)
# muliple bootstrap draws, keeping only coefficient estimates
bootstrap(rng::MersenneTwister, m::PHModel, iter::Int; kwargs...)
bootstrap(m::PHModel, iter::Int; kwargs...)

Returns:

If using bootstrap(m): a single bootstrap draw

If using bootstrap(m, 10) (e.g.): 10 bootstrap draws of the cumulative cause-specific risks at the end of follow up

using LSurvival, Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 500, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvivalResp(int, outt, d, ID.(id))    # specification with ID only
P = PHParms(X)

Mod = PHModel(R, P)
LSurvival._fit!(Mod, start=Mod.P._B, keepx=true, keepy=true)

# careful propogation of bootstrap sampling
idx, R2 = bootstrap(R)
P2 = bootstrap(idx, P)
Modb = PHModel(R2, P2)
LSurvival._fit!(Mod, start=Mod.P._B, keepx=true, keepy=true)

# convenience function for bootstrapping a model
Modc = bootstrap(Mod)
LSurvival._fit!(Modc, start=Modc.P._B);
Modc
Modc.P.X == nothing
Modc.R == nothing

Bootstrap Cox model coefficients

LSurvival._fit!(mb, keepx=true, keepy=true, start=[0.0, 0.0])

using LSurvival, Random
res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 200)
int = zeros(length(d)) # no late entry
X = hcat(z, x)

mainfit = fit(PHModel, X, int, outt, d .* (event .== 1), keepx=true, keepy=true)

function stddev_finite(x)
 n = length(x)
 mnx = sum(x)/n
 ret = sum((x .- mnx) .^ 2)
 ret /= n-1
 sqrt(ret)
end

# bootstrap standard error versus asymptotic
mb = bootstrap(MersenneTwister(123123), mainfit, 200)
## bootstrap standard error
[stddev_finite(mb[:,i]) for i in 1:2]
## asymptotic standard error
stderror(mainfit)

LSurvival.bootstrap — Method

Bootstrapping coefficients of a proportional hazards model

Signatures

# single bootstrap draw, keeping the entire object
bootstrap(rng::MersenneTwister, m::PHModel)
bootstrap(m::PHModel)
# muliple bootstrap draws, keeping only coefficient estimates
bootstrap(rng::MersenneTwister, m::PHModel, iter::Int; kwargs...)
bootstrap(m::PHModel, iter::Int; kwargs...)

Returns:

If using bootstrap(m): a single bootstrap draw

If using bootstrap(m, 10) (e.g.): 10 bootstrap draws of the cumulative cause-specific risks at the end of follow up

using LSurvival, Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 500, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvivalResp(int, outt, d, ID.(id))    # specification with ID only
P = PHParms(X)

Mod = PHModel(R, P)
LSurvival._fit!(Mod, start=Mod.P._B, keepx=true, keepy=true)

# careful propogation of bootstrap sampling
idx, R2 = bootstrap(R)
P2 = bootstrap(idx, P)
Modb = PHModel(R2, P2)
LSurvival._fit!(Mod, start=Mod.P._B, keepx=true, keepy=true)

# convenience function for bootstrapping a model
Modc = bootstrap(Mod)
LSurvival._fit!(Modc, start=Modc.P._B);
Modc
Modc.P.X == nothing
Modc.R == nothing

Bootstrap Cox model coefficients

LSurvival._fit!(mb, keepx=true, keepy=true, start=[0.0, 0.0])

using LSurvival, Random
res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 200)
int = zeros(length(d)) # no late entry
X = hcat(z, x)

mainfit = fit(PHModel, X, int, outt, d .* (event .== 1), keepx=true, keepy=true)

function stddev_finite(x)
 n = length(x)
 mnx = sum(x)/n
 ret = sum((x .- mnx) .^ 2)
 ret /= n-1
 sqrt(ret)
end

# bootstrap standard error versus asymptotic
mb = bootstrap(MersenneTwister(123123), mainfit, 200)
## bootstrap standard error
[stddev_finite(mb[:,i]) for i in 1:2]
## asymptotic standard error
stderror(mainfit)

LSurvival.bootstrap — Method

Bootstrap sampling of a proportional hazards predictor object

using LSurvival, Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvivalResp(int, outt, d, ID.(id))    # specification with ID only
P = PHParms(X)
idx, R2 = bootstrap(R)
P2 = bootstrap(idx, P)

Mod = PHModel(R2, P2)
LSurvival._fit!(Mod, start=Mod.P._B)

LSurvival.bootstrap — Method

Bootstrap methods for Aalen-Johansen cumulative risk estimator

Signatures

 # single bootstrap draw, keeping the entire object
 bootstrap(rng::MersenneTwister, m::AJSurv)
 bootstrap(m::AJSurv)

 # muliple bootstrap draws, keeping only coefficient estimates
 bootstrap(rng::MersenneTwister, m::AJSurv, iter::Int; kwargs...)
 bootstrap(m::AJSurv, iter::Int; kwargs...)

Returns:

If using bootstrap(m): a single bootstrap draw

If using bootstrap(m, 10) (e.g.): 10 bootstrap draws of the cumulative cause-specific risks at the end of follow up

using LSurvival
using Random

z, x, t, d, event, wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 100)
id = 1:length(x)
enter = zeros(length(t))

aj1 = aalen_johansen(enter, t, event, id=ID.(id), wts=wt)
aj2 = bootstrap(aj1, keepy=false);
ajboot = bootstrap(aj1, 10, keepy=false);
aj1

aj1.R
aj2.R

LSurvival.bootstrap — Method

Bootstrap methods for Kaplan-Meier survival curve estimator

Signatures

 # single bootstrap draw, keeping the entire object
 bootstrap(rng::MersenneTwister, m::KMSurv)
 bootstrap(m::KMSurv)

 # muliple bootstrap draws, keeping only coefficient estimates
 bootstrap(rng::MersenneTwister, m::KMSurv, iter::Int; kwargs...)
 bootstrap(m::KMSurv, iter::Int; kwargs...)

Returns:

If using bootstrap(m): a single bootstrap draw

If using bootstrap(m, 10) (e.g.): 10 bootstrap draws of the survival probability at the end of follow up

using LSurvival
using Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
wts = rand(length(d))

km1 = kaplan_meier(int, outt, d, id=ID.(id), wts=wts)
km2 = bootstrap(km1, keepy=false)
km3 = bootstrap(km1, 10, keepy=false)
km1

km1.R
km2.R

LSurvival.bootstrap — Method

Bootstrapping sampling of a competing risk survival response

Signatures

bootstrap(rng::MersenneTwister, R::T) where {T<:LSurvivalCompResp}
bootstrap(R::T) where {T<:LSurvivalCompResp}

z,x,t,d,event,weights =
LSurvival.dgm_comprisk(MersenneTwister(1212), 300)
enter = zeros(length(event))

# survival outcome:
R = LSurvivalCompResp(enter, t, event, weights, ID.(collect(1:length(t))))    # specification with ID only
bootstrap(R) # note that entire observations/clusters identified by id are kept

LSurvival.bootstrap — Method

Bootstrapping sampling of a survival response

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvivalResp(int, outt, d, ID.(id))    # specification with ID only

LSurvival.coxph — Method

Fit method for AbstractPH objects (Cox models)

Keyword arguments (used here, and passed on to internal structs)

ties "breslow" or "efron" (default)

wts observation weights

ties "breslow" or "efron" (default)

offset not currently used at all

fitargs arguments passed to other structs, which include

id cluster or individual level ID (defaults to a unique value for each row of data) see note below on ID

contrasts StatsModel style contrasts (dicts) that can be used for variable transformations/indicator variable creation (e.g. https://juliastats.org/StatsModels.jl/stable/contrasts/)

Signatures

fit(::Type{M},
X::AbstractMatrix,#{<:FP},
enter::AbstractVector{<:Real},
exit::AbstractVector{<:Real},
y::Union{AbstractVector{<:Real},BitVector}
;
ties ="breslow",
wts::AbstractVector{<:Real}      = similar(y, 0),
offset::AbstractVector{<:Real}   = similar(y, 0),
fitargs...) where {M<:AbstractPH}

```

coxph(f::FormulaTerm, data; kwargs...) ```

coxph(X, enter, exit, y, args...; kwargs...)

julia using LSurvival, Random z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000); enter = zeros(length(t)); X = hcat(x,rand(length(x))); m = fit(PHModel, X, enter, t, d, ties="efron") m2 = fit(PHModel, X, enter, t, d, ties="breslow") coeftable(m)

Note on use of id keyword

id is not needed in person-period structure data for standard estimates or confidence intervals

```@example using Random, LSurvival id, int, outt, dat = LSurvival.dgm(MersenneTwister(123123), 100, 100; afun = LSurvival.int_0) data = ( int = int, outt = outt, d = dat[:,4] .== 1, x = dat[:,1], z = dat[:,2] )

 f = @formula(Surv(int, outt,d)~x+z)
 coxph(f, data)

```

BUT, you must specify id to get appropriate robust variance and some other statistics.

Here is an example where the same data are presented in two different ways, which should yield identical statistics when used in Cox model. ```@example dat1 = ( time = [1,1,6,6,8,9], status = [1,0,1,1,0,1], x = [1,1,1,0,0,0] ) ft = coxph(@formula(Surv(time,status)~x),dat1) bic(ft) nobs(ft) dof_residual(ft)

lrtest is another one

stderror(ft) # model based stderror(ft, type="robust") # robust standard error, based on dfbeta residuals ft

now using "clustered" data with multiple observations per individual

dat1clust= ( id = [1,2,3,3,4,4,5,5,6,6], enter = [0,0,0,1,0,1,0,1,0,1], exit = [1,1,1,6,1,6,1,8,1,9], status = [1,0,0,1,0,1,0,0,0,1], x = [1,1,1,1,0,0,0,0,0,0] )

use the id parameter with the ID struct

ft2 = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust, id=ID.(dat1clust.id)) bic(ft2) # CORRECT nobs(ft2) # CORRECT dof_residual(ft2) # CORRECT

stderror(ft2) # model based (CORRECT) stderror(ft2, type="robust") # robust standard error, based on id level dfbeta residuals (CORRECT)

once robust SE is calculated, coefficient table uses the robust SE for confidence intervals and test statistics

ft2 # CORRECT (compare to ft object) ```

NOTE THE FOLLOWING IS INCORRECT because the id keyword is omitted

```@example ft2w = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust) bic(ft2w) # INCORRECT nobs(ft2w) # INCORRECT dof_residual(ft2w) # INCORRECT

stderror(ft2w) # model based (CORRECT) stderror(ft2w, type="robust") # robust variance (INCORRECT)

ft2w # the coefficient table now shows incorrect confidence intervals and test statistics

```

LSurvival.dgm — Method

Generating discrete survival data without competing risks

Signatures

dgm(rng::MersenneTwister, n::Int, maxT:Int; afun = int_0, yfun = yprob, lfun = lprob)

dgm(n::Int, maxT::Int; kwargs...)

Usage: dgm(rng, n, maxT;afun=int0, yfun=yprob, lfun=lprob) dgm(n, maxT;afun=int0, yfun=yprob, lfun=lprob)

Where afun, yfun, and lfun are all functions that take arguments v,l,a and output time-specific values of a, y, and l respectively Example:

expit(mu) =  inv(1.0+exp(-mu))

function aprob(v,l,a)
expit(-1.0 + 3*v + 2*l)
end
  
function lprob(v,l,a)
expit(-3 + 2*v + 0*l + 0*a)
end
  
function yprob(v,l,a)
expit(-3 + 2*v + 0*l + 2*a)
end
  # 10 individuals followed for up to 5 times
LSurvival.dgm(10, 5;afun=aprob, yfun=yprob, lfun=lprob)

LSurvival.dgm_comprisk — Method

Generating continuous survival data with competing risks

Signatures

dgm_comprisk(rng::MersenneTwister, n::Int)

dgm_comprisk(n::Int)

    - rng = random number generator    
    - n = sample size

Example:

using LSurvival
# 100 individuals with two competing events
z,x,t,d,event,weights = LSurvival.dgm_comprisk(100)
    

LSurvival.jackknife — Method

Obtain jackknife (leave-one-out) estimates from a Aalen-Johansen risk curve (risk at end of follow-up) by refitting the model n times

Signatures

jackknife(m::M;kwargs...) where {M<:AJSurv}

LSurvival.jackknife — Method

Obtain jackknife (leave-one-out) estimates from a Kaplan-Meier survival curve (survival at end of follow-up) by refitting the model n times

Signatures

jackknife(m::M;kwargs...) where {M<:KMSurv}

using LSurvival, Random, StatsBase

dat1 = (time = [1, 1, 6, 6, 8, 9], status = [1, 0, 1, 1, 0, 1], x = [1, 1, 1, 0, 0, 0])

dat1clust = (
  id = [1, 2, 3, 3, 4, 4, 5, 5, 6, 6],
  enter = [0, 0, 0, 1, 0, 1, 0, 1, 0, 1],
  exit = [1, 1, 1, 6, 1, 6, 1, 8, 1, 9],
  status = [1, 0, 0, 1, 0, 1, 0, 0, 0, 1],
  x = [1, 1, 1, 1, 0, 0, 0, 0, 0, 0],
)

m = kaplan_meier(dat1.time, dat1.status)
a = aalen_johansen(dat1.time, dat1.status)
mc = kaplan_meier(dat1clust.enter, dat1clust.exit, dat1clust.status, id=ID.(dat1clust.id))
ac = aalen_johansen(dat1clust.enter, dat1clust.exit, dat1clust.status, id=ID.(dat1clust.id))
jk = jackknife(m);
jkc = jackknife(mc);
jka = jackknife(a);
bs = bootstrap(mc, 100);
std(bs[:,1])
stderror(m, type="jackknife")
stderror(mc, type="jackknife")
@assert jk == jkc

LSurvival.jackknife — Method

Obtain jackknife (leave-one-out) estimates from a Cox model by refitting the model n times

using LSurvival, Random, StatsBase
id, int, outt, data =
LSurvival.dgm(MersenneTwister(112), 100, 10; afun = LSurvival.int_0)
data[:, 1] = round.(data[:, 1], digits = 3)
d, X = data[:, 4], data[:, 1:3]
wt = rand(length(d))
wt ./= (sum(wt) / length(wt))
m = coxph(X,int, outt,d, wts=wt, id=ID.(id))

jk = jackknife(m);
bs = bootstrap(MersenneTwister(12321), m, 1000);
N = nobs(m)
#comparing estimate with jackknife estimate with bootstrap mean
hcat(coef(m), mean(jk, dims=1)[1,:], mean(bs, dims=1)[1,:])
semb = stderror(m)
sebs = std(bs, dims=1)
sero = stderror(m, type="robust")
sejk = stderror(m, type="jackknife")
sejk_manual = std(jk, dims=1, corrected=false) .* sqrt(N-1)

sqrt.(diag(LSurvival.jackknife_vcov(m)))

hcat(semb, sebs[1,:], sejk, sejk_manual[1,:], sero)

dat1 = (time = [1, 1, 6, 6, 8, 9], status = [1, 0, 1, 1, 0, 1], x = [1, 1, 1, 0, 0, 0])
dat1clust = (
  id = [1, 2, 3, 3, 4, 4, 5, 5, 6, 6],
  enter = [0, 0, 0, 1, 0, 1, 0, 1, 0, 1],
  exit = [1, 1, 1, 6, 1, 6, 1, 8, 1, 9],
  status = [1, 0, 0, 1, 0, 1, 0, 0, 0, 1],
  x = [1, 1, 1, 1, 0, 0, 0, 0, 0, 0],
)

m = coxph(@formula(Surv(time, status)~x),dat1)
mc = coxph(@formula(Surv(enter, exit, status)~x),dat1clust, id=ID.(dat1clust.id))
jk = jackknife(m);
jkc = jackknife(mc);
bs = bootstrap(mc, 100);
std(bs[:,1])
stderror(m, type="jackknife")
stderror(mc, type="jackknife")
@assert jk == jkc

LSurvival.kaplan_meier — Method

Kaplan-Meier estimator for cumulative conditional risk

Signatures

StatsBase.fit!(m::T; kwargs...) where {T<:AbstractNPSurv}

kaplan_meier(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; kwargs...)

Keyword arguments

wts::Vector{<:Real} = similar(enter, 0); vector of case weights (or zero length vector) for each observation

id::Vector{<:AbstractLSurvivalID} = [ID(i) for i in eachindex(y)]; Vector of AbstractSurvID objects denoting observations that form a single unit (used in bootstrap and jackknife methods)

atol = 0.00000001; absolute tolerance for defining tied event times

censval = 0; value of the outcome to be considered a censored event

keepy = true; keep the outcome vector after fitting (may save memory with large datasets)

eps = 0.00000001; deprecated (replaced by atol)

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)

or, equivalently:

kaplan_meier(enter, t, d, wts=wt)

LSurvival.lgh_breslow! — Method

Update the partial likelihood, gradient and Hessian values from a Cox model fit (used during fitting, not generally useful for users).

Uses Breslow's partial likelihood.

Updates over all observations

Signature

lgh_breslow!(m::M, j, caseidx, risksetidx) where {M<:AbstractPH}

LSurvival.lgh_efron! — Method

Update the partial likelihood, gradient and Hessian values from a Cox model fit (used during fitting, not generally useful for users).

Uses Efron's partial likelihood.

Updates over all observations

Signature

lgh_efron!(m::M, j, caseidx, risksetidx) where {M<:AbstractPH}

LSurvival.risk_from_coxphmodels — Method

Survival curve estimation using multiple cox models

Signatures

```julia riskfromcoxphmodels(fitlist::Vector{T}, args...; kwargs...) where {T<:PHModel}

fit(::Type{M}, fitlist::Vector{T}, ; fitargs...) where {M<:PHSurv,T<:PHModel} ```

Optional keywords

coef_vectors = nothing(default) or vector of coefficient vectors from the cox models [will default to the coefficients from fitlist models]

pred_profile = nothing(default) or vector of specific predictor values of the same length as the coef_vectors[1]

```@example using LSurvival using Random

event variable is coded 0[referent],1,2

z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000); enter = zeros(length(t));

ft1 = coxph(hcat(x,z), enter, t, (event .== 1)) nidx = findall(event .!= 1) ft2 = coxph(hcat(x,z)[nidx,:], enter[nidx], t[nidx], (event[nidx] .== 2))

risk at referent levels of x and z

riskfromcoxphmodels([ft1,ft2])

risk at average levels of x and z

mnx = sum(x)/length(x) mnz = sum(z)/length(z)

equivalent

fit(PHSurv, [ft1,ft2], predprofile=[mnx,mnz]) riskfromcoxphmodels([ft1,ft2], predprofile=[mnx,mnz]) ```

RecipesBase.apply_recipe — Method

Recipe for aalen-johansen risk curve

    using Plots, LSurvival
    res = z, x, outt, d, event, weights = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
    int = zeros(length(d)) # no late entry
    
        c = fit(AJSurv, int, outt, event)
        #risk2 = aalen_johansen(int, outt, event)
        plot(c)

RecipesBase.apply_recipe — Method

Recipe for kaplan meier curve

    using Plots, LSurvival
dat4 = (
    id = [1, 1, 2, 2, 2, 3, 4, 5, 5, 6],
    enter = [1, 2, 5, 4, 6, 7, 3, 6, 8, 0],
    exit = [2, 5, 6, 7, 8, 9, 6, 8, 14, 9],
    status = [0, 1, 0, 0, 1, 0, 1, 0, 0, 1],
    x = [0.1, 0.1, 1.5, 1.5, 1.5, 0, 0, 0, 0, 3],
    z = [1, 1, 0, 0, 0, 0, 0, 1, 1, 0],
    w = [0, 0, 0, 0, 0, 1, 1, 1, 1, 0],
)
R = LSurvivalResp(dat4.enter, dat4.exit, dat4.status)
    k = kaplan_meier(dat4.enter, dat4.exit, dat4.status)
    plot(k)

RecipesBase.apply_recipe — Method

Plotting baseline hazard for a Cox model

using Plots, LSurvival
dat2 = (
    enter = [1, 2, 5, 2, 1, 7, 3, 4, 8, 8],
    exit = [2, 3, 6, 7, 8, 9, 9, 9, 14, 17],
    status = [1, 1, 1, 1, 1, 1, 1, 0, 0, 0],
    x = [1, 0, 0, 1, 0, 1, 1, 1, 0, 0],
)
fte = coxph(@formula(Surv(enter, exit, status)~x), dat2, maxiter=0)
ftb = coxph(@formula(Surv(enter, exit, status)~x), dat2, ties="breslow", maxiter=0)

plot(fte, label="Efron")
plot!(ftb, label="Breslow")

RecipesBase.apply_recipe — Method

using Plots, LSurvival
dat2 = (
    enter = [1, 2, 5, 2, 1, 7, 3, 4, 8, 8],
    exit = [2, 3, 6, 7, 8, 9, 9, 9, 14, 17],
    status = [1, 1, 1, 1, 1, 1, 1, 0, 0, 0],
    x = [1, 0, 0, 1, 0, 1, 1, 1, 0, 0],
)
fte = coxph(@formula(Surv(enter, exit, status)~x), dat2)

coxdx(fte)

RecipesBase.apply_recipe — Method

using Plots, LSurvival
dat2 = (
    enter = [1, 2, 5, 2, 1, 7, 3, 4, 8, 8],
    exit = [2, 3, 6, 7, 8, 9, 9, 9, 14, 17],
    status = [1, 1, 1, 1, 1, 1, 1, 0, 0, 0],
    x = [1, 0, 0, 1, 0, 1, 1, 1, 0, 0],
)
fte = coxph(@formula(Surv(enter, exit, status)~x), dat2)

coxinfluence(fte, type="jackknife", par=1)
coxinfluence!(fte, type="dfbeta", color=:red, par=1)

RecipesBase.apply_recipe — Method

using Plots, LSurvival

dat4 = (
    id = [1, 1, 2, 2, 2, 3, 4, 5, 5, 6],
    enter = [1, 2, 5, 4, 6, 7, 3, 6, 8, 0],
    exit = [2, 5, 6, 7, 8, 9, 6, 8, 14, 9],
    status = [0, 1, 0, 0, 1, 0, 1, 0, 0, 1],
    x = [0.1, 0.1, 1.5, 1.5, 1.5, 0, 0, 0, 0, 3],
    z = [1, 1, 0, 0, 0, 0, 0, 1, 1, 0],
    w = [0, 0, 0, 0, 0, 1, 1, 1, 1, 0],
)
R = LSurvivalResp(dat4.enter, dat4.exit, dat4.status)
plot([[R.enter[i], R.exit[i]] for i in eachindex(R.enter)], [[i, i] for i in values(R.id)])

RecipesBase.apply_recipe — Method

Recipe for cox-model based risk curves

    using Plots, LSurvival, Random, StatsBase
    res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
    X = hcat(z, x)
    int = zeros(length(d)) # no late entry
    ft1 = fit(PHModel, X, int, outt, d .* (event .== 1), wts=wts)
    ft2 = fit(PHModel, X, int, outt, d .* (event .== 2), wts=wts)
    c = risk_from_coxphmodels([ft1, ft2], pred_profile = mean(X, dims=1))
    
    plot(c)

StatsAPI.confint — Method

Greenwood's formula for variance and confidence intervals of a Aalen-Johansen risk function

Signatures:

StatsBase.stderror(m::AJSurv)

StatsBase.confint(m:AJSurv; level=0.95, method="normal")

Keyword arguments

method

"normal" normality-based confidence intervals

"lognlog" log(-log(S(t))) based confidence intervals

res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
int = zeros(length(d)) # no late entry
m = fit(AJSurv, int, outt, event)
stderror(m)
confint(m, level=0.95)

StatsAPI.confint — Method

Greenwood's formula for variance and confidence intervals of a Kaplan-Meier survival curve

Signatures:

StatsBase.stderror(m::KMSurv)

StatsBase.confint(m:KMSurv; level=0.95, method="normal")

Keyword arguments

method:

"normal" normality-based confidence intervals

"lognlog" log(-log(S(t))) based confidence intervals

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)
stderror(m)
confint(m, method="normal")
confint(m, method="lognlog") # log-log transformation

StatsAPI.confint — Method

using LSurvival
 dat1= (
   time = [1,1,6,6,8,9],
   status = [1,0,1,1,0,1],
   x = [1,1,1,0,0,0]
 )

 ft = coxph(@formula(Surv(time, status) ~ x),dat1, keepx=true)
 # model-based variance
 confint(ft)

 # robust variance
 confint(ft, type="robust")

for cluster confidence intervals

```julia dat1clust= ( id = [1,2,3,3,4,4,5,5,6,6], enter = [0,0,0,1,0,1,0,1,0,1], exit = [1,1,1,6,1,6,1,8,1,9], status = [1,0,0,1,0,1,0,0,0,1], x = [1,1,1,1,0,0,0,0,0,0] )

ft2 = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust, id=ID.(dat1clust.id), keepx=true)

model-based variance

confint(ft2)

robust variance

confint(ft2, type="robust") ````

StatsAPI.fit! — Method

Kaplan-Meier estimator for cumulative conditional risk

Signatures

StatsBase.fit!(m::T; kwargs...) where {T<:AbstractNPSurv}

kaplan_meier(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; kwargs...)

Keyword arguments

wts::Vector{<:Real} = similar(enter, 0); vector of case weights (or zero length vector) for each observation

id::Vector{<:AbstractLSurvivalID} = [ID(i) for i in eachindex(y)]; Vector of AbstractSurvID objects denoting observations that form a single unit (used in bootstrap and jackknife methods)

atol = 0.00000001; absolute tolerance for defining tied event times

censval = 0; value of the outcome to be considered a censored event

keepy = true; keep the outcome vector after fitting (may save memory with large datasets)

eps = 0.00000001; deprecated (replaced by atol)

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)

or, equivalently:

kaplan_meier(enter, t, d, wts=wt)

Aalen-Johansen estimator for cumulative cause-specific risk (in the presence of competing events)

Signatures

 StatsBase.fit!(m::T; kwargs...) where {T<:AbstractNPSurv}

 aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; kwargs...)

Keyword arguments

wts::Vector{<:Real} = similar(enter, 0); vector of case weights (or zero length vector) for each observation

id::Vector{<:AbstractLSurvivalID} = [ID(i) for i in eachindex(y)]; Vector of AbstractSurvID objects denoting observations that form a single unit (used in bootstrap and jackknife methods)

atol = 0.00000001; absolute tolerance for defining tied event times

keepy = true; keep the outcome vector after fitting (may save memory with large datasets)

eps = 0.00000001; deprecated (replaced by atol)

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter, t, event, wts=wt)

Survival curve estimation using multiple cox models

Signatures

```julia riskfromcoxphmodels(fitlist::Vector{T}, args...; kwargs...) where {T<:PHModel}

fit(::Type{M}, fitlist::Vector{T}, ; fitargs...) where {M<:PHSurv,T<:PHModel} ```

Optional keywords

coef_vectors = nothing(default) or vector of coefficient vectors from the cox models [will default to the coefficients from fitlist models]

pred_profile = nothing(default) or vector of specific predictor values of the same length as the coef_vectors[1]

```@example using LSurvival using Random

event variable is coded 0[referent],1,2

z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000); enter = zeros(length(t));

ft1 = coxph(hcat(x,z), enter, t, (event .== 1)) nidx = findall(event .!= 1) ft2 = coxph(hcat(x,z)[nidx,:], enter[nidx], t[nidx], (event[nidx] .== 2))

risk at referent levels of x and z

riskfromcoxphmodels([ft1,ft2])

risk at average levels of x and z

mnx = sum(x)/length(x) mnz = sum(z)/length(z)

equivalent

fit(PHSurv, [ft1,ft2], predprofile=[mnx,mnz]) riskfromcoxphmodels([ft1,ft2], predprofile=[mnx,mnz]) ```

StatsAPI.fit — Method

Survival curve estimation using multiple cox models

Signatures

```julia riskfromcoxphmodels(fitlist::Vector{T}, args...; kwargs...) where {T<:PHModel}

fit(::Type{M}, fitlist::Vector{T}, ; fitargs...) where {M<:PHSurv,T<:PHModel} ```

Optional keywords

coef_vectors = nothing(default) or vector of coefficient vectors from the cox models [will default to the coefficients from fitlist models]

pred_profile = nothing(default) or vector of specific predictor values of the same length as the coef_vectors[1]

```@example using LSurvival using Random

event variable is coded 0[referent],1,2

z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000); enter = zeros(length(t));

ft1 = coxph(hcat(x,z), enter, t, (event .== 1)) nidx = findall(event .!= 1) ft2 = coxph(hcat(x,z)[nidx,:], enter[nidx], t[nidx], (event[nidx] .== 2))

risk at referent levels of x and z

riskfromcoxphmodels([ft1,ft2])

risk at average levels of x and z

mnx = sum(x)/length(x) mnz = sum(z)/length(z)

equivalent

fit(PHSurv, [ft1,ft2], predprofile=[mnx,mnz]) riskfromcoxphmodels([ft1,ft2], predprofile=[mnx,mnz]) ```

StatsAPI.fit — Method

Fit method for AbstractPH objects (Cox models)

Keyword arguments (used here, and passed on to internal structs)

ties "breslow" or "efron" (default)

wts observation weights

ties "breslow" or "efron" (default)

offset not currently used at all

fitargs arguments passed to other structs, which include

id cluster or individual level ID (defaults to a unique value for each row of data) see note below on ID

contrasts StatsModel style contrasts (dicts) that can be used for variable transformations/indicator variable creation (e.g. https://juliastats.org/StatsModels.jl/stable/contrasts/)

Signatures

fit(::Type{M},
X::AbstractMatrix,#{<:FP},
enter::AbstractVector{<:Real},
exit::AbstractVector{<:Real},
y::Union{AbstractVector{<:Real},BitVector}
;
ties ="breslow",
wts::AbstractVector{<:Real}      = similar(y, 0),
offset::AbstractVector{<:Real}   = similar(y, 0),
fitargs...) where {M<:AbstractPH}

```

coxph(f::FormulaTerm, data; kwargs...) ```

coxph(X, enter, exit, y, args...; kwargs...)

julia using LSurvival, Random z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000); enter = zeros(length(t)); X = hcat(x,rand(length(x))); m = fit(PHModel, X, enter, t, d, ties="efron") m2 = fit(PHModel, X, enter, t, d, ties="breslow") coeftable(m)

Note on use of id keyword

id is not needed in person-period structure data for standard estimates or confidence intervals

```@example using Random, LSurvival id, int, outt, dat = LSurvival.dgm(MersenneTwister(123123), 100, 100; afun = LSurvival.int_0) data = ( int = int, outt = outt, d = dat[:,4] .== 1, x = dat[:,1], z = dat[:,2] )

 f = @formula(Surv(int, outt,d)~x+z)
 coxph(f, data)

```

BUT, you must specify id to get appropriate robust variance and some other statistics.

Here is an example where the same data are presented in two different ways, which should yield identical statistics when used in Cox model. ```@example dat1 = ( time = [1,1,6,6,8,9], status = [1,0,1,1,0,1], x = [1,1,1,0,0,0] ) ft = coxph(@formula(Surv(time,status)~x),dat1) bic(ft) nobs(ft) dof_residual(ft)

lrtest is another one

stderror(ft) # model based stderror(ft, type="robust") # robust standard error, based on dfbeta residuals ft

now using "clustered" data with multiple observations per individual

dat1clust= ( id = [1,2,3,3,4,4,5,5,6,6], enter = [0,0,0,1,0,1,0,1,0,1], exit = [1,1,1,6,1,6,1,8,1,9], status = [1,0,0,1,0,1,0,0,0,1], x = [1,1,1,1,0,0,0,0,0,0] )

use the id parameter with the ID struct

ft2 = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust, id=ID.(dat1clust.id)) bic(ft2) # CORRECT nobs(ft2) # CORRECT dof_residual(ft2) # CORRECT

stderror(ft2) # model based (CORRECT) stderror(ft2, type="robust") # robust standard error, based on id level dfbeta residuals (CORRECT)

once robust SE is calculated, coefficient table uses the robust SE for confidence intervals and test statistics

ft2 # CORRECT (compare to ft object) ```

NOTE THE FOLLOWING IS INCORRECT because the id keyword is omitted

```@example ft2w = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust) bic(ft2w) # INCORRECT nobs(ft2w) # INCORRECT dof_residual(ft2w) # INCORRECT

stderror(ft2w) # model based (CORRECT) stderror(ft2w, type="robust") # robust variance (INCORRECT)

ft2w # the coefficient table now shows incorrect confidence intervals and test statistics

```

StatsAPI.fit — Method

Aalen-Johansen estimator for cumulative cause-specific risk (in the presence of competing events)

Signatures

 StatsBase.fit!(m::T; kwargs...) where {T<:AbstractNPSurv}

 aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; kwargs...)

Keyword arguments

wts::Vector{<:Real} = similar(enter, 0); vector of case weights (or zero length vector) for each observation

id::Vector{<:AbstractLSurvivalID} = [ID(i) for i in eachindex(y)]; Vector of AbstractSurvID objects denoting observations that form a single unit (used in bootstrap and jackknife methods)

atol = 0.00000001; absolute tolerance for defining tied event times

keepy = true; keep the outcome vector after fitting (may save memory with large datasets)

eps = 0.00000001; deprecated (replaced by atol)

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter, t, event, wts=wt)

StatsAPI.fit — Method

Kaplan-Meier estimator for cumulative conditional risk

Signatures

StatsBase.fit!(m::T; kwargs...) where {T<:AbstractNPSurv}

kaplan_meier(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; kwargs...)

Keyword arguments

wts::Vector{<:Real} = similar(enter, 0); vector of case weights (or zero length vector) for each observation

id::Vector{<:AbstractLSurvivalID} = [ID(i) for i in eachindex(y)]; Vector of AbstractSurvID objects denoting observations that form a single unit (used in bootstrap and jackknife methods)

atol = 0.00000001; absolute tolerance for defining tied event times

censval = 0; value of the outcome to be considered a censored event

keepy = true; keep the outcome vector after fitting (may save memory with large datasets)

eps = 0.00000001; deprecated (replaced by atol)

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)

or, equivalently:

kaplan_meier(enter, t, d, wts=wt)

StatsAPI.loglikelihood — Method

Maximum log partial likelihood for a fitted AbstractPH model Efron or Breslow (depending on the ties` parameter)

StatsAPI.nullloglikelihood — Method

Null log-partial likelihood for a fitted AbstractPH model Efron or Breslow (depending on the ties` parameter)

Note: this is just the log partial likelihood at the initial values of the model, which default to 0. If initial values are non-null, then this function no longer validly returns the null log-partial likelihood.

StatsAPI.residuals — Method

####################################################################

Cox proportional hazards model residuals:

Signature

julia residuals(m::M; type = "martingale") where {M<:PHModel} where type is one of

martingale

schoenfeld

score

dfbeta

scaled_schoenfeld

Residuals from the residuals function are designed to exactly emulate those from the survival package in R. Currently, they are validated for single observation data (e.g. one data row per individual).

####################################################################

Martingale residuals: Observed versus expected

# example from https://cran.r-project.org/web/packages/survival/vignettes/validate.pdf
# by Terry Therneau

dat1 = (
  time = [1,1,6,6,8,9],
  status = [1,0,1,1,0,1],
  x = [1,1,1,0,0,0]
)

# Nelson-Aalen type estimator for Breslow partial likelihood
ft = coxph(@formula(Surv(time,status)~x),dat1, keepx=true, keepy=true, ties="breslow")
residuals(ft, type="martingale")

dat1 = (
  time = [1,1,6,6,8,9],
  status = [1,0,1,1,0,1],
  x = [1,1,1,0,0,0]
)

# Fleming-Harrington type estimator for Efron partial likelihood
ft = coxph(@formula(Surv(time,status)~x),dat1, keepx=true, keepy=true, ties="efron")
residuals(ft, type="martingale")

####################################################################

Score residuals: Per observation contribution to score function

using LSurvival
dat1 = (
  time = [1,1,6,6,8,9],
  status = [1,0,1,1,0,1],
  x = [1,1,1,0,0,0]
)
ft = coxph(@formula(Surv(time,status)~x),dat1, keepx=true, keepy=true, ties="breslow")
S = residuals(ft, type="score")[:]
ft = coxph(@formula(Surv(time,status)~x),dat1, keepx=true, keepy=true, ties="efron", maxiter=0)
S = residuals(ft, type="score")[:]

####################################################################

Schoenfeld residuals: Per time contribution to score function

```julia using LSurvival dat1 = ( time = [1,1,6,6,8,9], status = [1,0,1,1,0,1], x = [1,1,1,0,0,0] ) ft = coxph(@formula(Surv(time,status)~x),dat1, keepx=true, keepy=true, ties="breslow", maxiter=0)

X = ft.P.X M = residuals(ft, type="martingale") S = residuals(ft, type="schoenfeld")[:] ```

####################################################################

dfbeta residuals: influence of individual observations on each parameter

```@example using LSurvival dat1 = ( time = [1,1,6,6,8,9], status = [1,0,1,1,0,1], x = [1,1,1,0,0,0] )

ft = coxph(@formula(Surv(time,status)~x),dat1, ties="breslow") residuals(ft, type="dfbeta")

can also calculate from score residuals and Hessian matrix

L = residuals(ft, type="score") # n X p H = ft.P._hess # p X p dfbeta = L*inv(H) robVar = dfbeta'dfbeta sqrt(robVar)

```

using the id keyword argument

see help for LSurvival.vcov for what happens when id keyword is not used

```@example dat1clust= ( id = [1,2,3,3,4,4,5,5,6,6], enter = [0,0,0,1,0,1,0,1,0,1], exit = [1,1,1,6,1,6,1,8,1,9], status = [1,0,0,1,0,1,0,0,0,1], x = [1,1,1,1,0,0,0,0,0,0] )

ft2 = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust, id=ID.(dat1clust.id), ties="breslow")

note these are still on the observation level (not the id level)!

residuals(ft2, type="dfbeta")

getting id level dfbeta residuals

dfbeta = residuals(ft2, type="dfbeta") id = values(ft2.R.id) D = reduce(vcat, [sum(dfbeta[findall(id .== i),:], dims=1) for i in unique(id)]) D'D vcov(ft, type="robust") vcov(ft2, type="robust") ```

####################################################################

jackknife residuals: influence of individual observations on each parameter using leave-one-out estimates

note there are other definitions of jackknife residuals See Chapter 7.1 of "Extending the Cox Model" by Therneau and Grambsch for an example of the type of jackknife residuals used here

Jackknife residuals $r_i$ for $i \in 1:n$ are given as the difference between the maximum partial likelihood estimate and the jackknife estimates for each observation

\[r_i = \hat\beta - \hat\beta_{(-i)}\]

where $\beta_{(-i)}$ is the maximum partial likelihood estimate of the log-hazard ratio vector obtained from a dataset in which observations belonging to individual i are removed

```@example using LSurvival dat1 = ( time = [1,1,6,6,8,9], status = [1,0,1,1,0,1], x = [1,1,1,0,0,0] )

ft = coxph(@formula(Surv(time,status)~x),dat1, ties="breslow") #jackknife(ft) residuals(ft, type="jackknife")

```

StatsAPI.stderror — Method

Greenwood's formula for variance and confidence intervals of a Aalen-Johansen risk function

Signatures:

StatsBase.stderror(m::AJSurv)

StatsBase.confint(m:AJSurv; level=0.95, method="normal")

Keyword arguments

method

"normal" normality-based confidence intervals

"lognlog" log(-log(S(t))) based confidence intervals

res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
int = zeros(length(d)) # no late entry
m = fit(AJSurv, int, outt, event)
stderror(m)
confint(m, level=0.95)

StatsAPI.stderror — Method

Greenwood's formula for variance and confidence intervals of a Kaplan-Meier survival curve

Signatures:

StatsBase.stderror(m::KMSurv)

StatsBase.confint(m:KMSurv; level=0.95, method="normal")

Keyword arguments

method:

"normal" normality-based confidence intervals

"lognlog" log(-log(S(t))) based confidence intervals

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)
stderror(m)
confint(m, method="normal")
confint(m, method="lognlog") # log-log transformation

StatsAPI.vcov — Method

Covariance matrix for Cox proportional hazards models

Keyword arguments

type nothing or "robust": determines whether model based or robust (dfbeta based) variance is returned.

See ?residuals for info on dfbeta residuals

using LSurvival
dat1 = (
  time = [1,1,6,6,8,9],
  status = [1,0,1,1,0,1],
  x = [1,1,1,0,0,0]
)
ft = coxph(@formula(Surv(time,status)~x),dat1, id=ID.(collect(1:6)))

vcov(ft)                   # model based
vcov(ft, type="robust")    # robust variance, based on dfbeta residuals
# once robust SE is calculated, coefficient table uses the robust SE for confidence intervals and test statistics
ft

cluster robust standard errors using the id keyword argument

dat1clust= (
  id = [1,2,3,3,4,4,5,5,6,6],
  enter = [0,0,0,1,0,1,0,1,0,1],
  exit = [1,1,1,6,1,6,1,8,1,9],
  status = [1,0,0,1,0,1,0,0,0,1],
  x = [1,1,1,1,0,0,0,0,0,0]
)

ft2 = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust, id=ID.(dat1clust.id))

vcov(ft2)                     # model based
vcov(ft2, type="robust")       # robust variance, based on dfbeta residuals
stderror(ft2, type="robust")   # robust variance, based on dfbeta residuals
confint(ft2, type="robust")    # robust variance, based on dfbeta residuals
nobs(ft2)                     # id argument yields correct value of number of independent observations
# once robust SE is calculated, coefficient table uses the robust SE for confidence intervals and test statistics
ft2 

NOTE THE FOLLOWING IS INCORRECT because the id keyword is omitted

ft2w = coxph(@formula(Surv(enter, exit, status) ~ x),dat1clust)

vcov(ft2w)                   # model based (CORRECT)
vcov(ft2w, type="robust")    # robust variance (INCORRECT)
nobs(ft2w)

ft2w