LSurvival.

Survival analysis functions in Julia for time-to-event outcomes that can include:

Loss-to-follow-up/right censoring

Late entry/left truncation (not available in the Survival.jl package)

"Person-period" data structures (not available in the Survival.jl package)

Observation weights (not available in the Survival.jl package)

Competing risks (not available in the Survival.jl package)

Capabilities include estimators for

Kaplan-Meier non-parametric conditional risk functions

Aalen-Johansen non-parametric cause-specific unconditional risk functions

Cox proportional hazards model (Efron's or Breslow's methods for ties)

Convenience functions enable:

Non-parametric bootstrapping, cluster-bootstrapping

Estimating baseline hazards from a Cox Model

Estimating cause-specific risk from an exhaustive set of Cox models for competing risk outcomes

Simple simulation of competing and non-competing survival events

Plans to include:

cluster robust variance estimation (without bootstrapping)

Report issues here

Installation

using Pkg; Pkg.add(url = "https://github.com/alexpkeil1/LSurvival.jl")

Quick examples

using Random, LSurvival, Distributions, LinearAlgebra

# generate some data
expit(mu) = inv(1.0 + exp(-mu))

function int_nc(v, l, a)
    expit(-1.0 + 3 * v + 2 * l)
end

function int_0(v, l, a)
    0.1
end

function lprob(v, l, a)
    expit(-3 + 2 * v + 0 * l + 0 * a)
end

function yprob(v, l, a)
    expit(-3 + 2 * v + 0 * l + 2 * a)
end

function dgm(rng, n, maxT; regimefun = int_0)
    V = rand(rng, n)
    LAY = Array{Float64,2}(undef, n * maxT, 4)
    keep = ones(Bool, n * maxT)
    id = sort(reduce(vcat, fill(collect(1:n), maxT)))
    time = (reduce(vcat, fill(collect(1:maxT), n)))
    for i = 1:n
        v = V[i]
        l = 0
        a = 0
        lkeep = true
        for t = 1:maxT
            currIDX = (i - 1) * maxT + t
            l = lprob(v, l, a) > rand(rng) ? 1 : 0
            a = regimefun(v, l, a) > rand(rng) ? 1 : 0
            y = yprob(v, l, a) > rand(rng) ? 1 : 0
            LAY[currIDX, :] .= [v, l, a, y]
            keep[currIDX] = lkeep
            lkeep = (!lkeep || (y == 1)) ? false : true
        end
    end
    id[findall(keep)], time[findall(keep)] .- 1, time[findall(keep)], LAY[findall(keep), :]
end

id, int, outt, data = dgm(MersenneTwister(), 1000, 10; regimefun = int_0)
data[:, 1] = round.(data[:, 1], digits = 3)
d, X = data[:, 4], data[:, 1:3]
wt = rand(length(d))

# Cox model
m = fit(PHModel, X, int, outt, d, ties = "breslow", wts = wt)
m2 = fit(PHModel, X, int, outt, d, ties = "efron", wts = wt)
#equivalent
m2b = coxph(X, int, outt, d, ties = "efron", wts = wt)

# Kaplan-Meier estimator of the cumulative risk/survival
res = kaplan_meier(int, outt, d)

# Competing risk analysis with Aalen-Johansen estimator of the cumulative risk/survival

function dgm_comprisk(; n = 100, rng = MersenneTwister())
    z = rand(rng, n) .* 5
    x = rand(rng, n) .* 5
    dt1 = Weibull.(fill(0.75, n), inv.(exp.(-x .- z)))
    dt2 = Weibull.(fill(0.75, n), inv.(exp.(-x .- z)))
    t01 = rand.(rng, dt1)
    t02 = rand.(rng, dt2)
    t0 = min.(t01, t02)
    t = Array{Float64,1}(undef, n)
    for i = 1:n
        t[i] = t0[i] > 1.0 ? 1.0 : t0[i]
    end
    d = (t .== t0)
    event = (t .== t01) .+ 2.0 .* (t .== t02)
    wtu = rand(rng, n) .* 5.0
    wt = wtu ./ mean(wtu)
    reshape(round.(z, digits = 4), (n, 1)),
    reshape(round.(x, digits = 4), (n, 1)),
    round.(t, digits = 4),
    d,
    event,
    round.(wt, digits = 4)
end

z, x, t, d, event, wt = dgm_comprisk(; n = 100, rng = MersenneTwister())
X = hcat(x,z)
enter = t .* rand(100)*0.02 # create some fake entry times

res = aalen_johansen(enter, t, event; wts = wt)
fit1 = fit(PHModel, X, enter, t, (event .== 1), ties = "breslow", wts = wt)
fit2 = fit(PHModel, X, enter, t, (event .== 1), ties = "efron", wts = wt)
risk_from_coxphmodels([fit1, fit2])

# this approach operates on left censored outcomes (which operate in the background in model fitting)
LSurvResp(enter, t, d)
LSurvCompResp(enter, t, event)

# can use the ID type to refer to units with multiple observations
id, int, outt, data = dgm(MersenneTwister(), 1000, 10; regimefun = int_0)
LSurvResp(int, outt, data[:,4], ID.(id))

Index of functions

LSurvival.AbstractLSurvParms

LSurvival.AbstractLSurvResp

LSurvival.AbstractNPSurv

LSurvival.AbstractPH

LSurvival.ID

LSurvival.LSurvCompResp

LSurvival.LSurvResp

LSurvival.PHModel

LSurvival.PHModel

LSurvival.PHModel

LSurvival.PHModel

LSurvival.PHModel

LSurvival.PHModel

LSurvival.PHSurv

LSurvival.PHSurv

LSurvival.PHSurv

LSurvival.Strata

LSurvival._update_PHParms!

LSurvival.aalen_johansen

LSurvival.aj

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.bootstrap

LSurvival.ci_from_coxmodels

LSurvival.coxmodel

LSurvival.coxph

LSurvival.coxph

LSurvival.dgm

LSurvival.dgm_comprisk

LSurvival.e_yearsoflifelost

LSurvival.kaplan_meier

LSurvival.km

LSurvival.km

LSurvival.lgh!

LSurvival.lgh_breslow!

LSurvival.lgh_breslow!

LSurvival.lgh_efron!

LSurvival.lgh_efron!

LSurvival.risk_from_coxphmodels

LSurvival.subdistribution_hazard_cuminc

StatsAPI.confint

StatsAPI.confint

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit!

StatsAPI.stderror

StatsAPI.stderror

Function help

LSurvival.AbstractLSurvParms — Type

AbstractLsurvParms

Abstract type representing a model predictors and coefficient parameters

LSurvival.AbstractLSurvResp — Type

AbstractLsurvResp

Abstract type representing a model response vector

LSurvival.AbstractNPSurv — Type

Abstract type for non-parametric survival models, including Kaplan-Meier, Aalen Johansen, and Cox-model based estimates of survival using an Aalen-Johansen-like estimator

LSurvival.AbstractPH — Type

Abstract type for proportional hazards models

LSurvival.ID — Type

Type for identifying individuals in survival outcomes.

Accepts any Number or String

[ID(i) for i in 1:10]

Used for the id argument in

Outcome types: LSurvResp, LSurvCompResp

Model types: PHModel, KMRisk, AJRisk

LSurvival.LSurvCompResp — Type

Outcome type for competing risk survival outcomes subject to left truncation and right censoring

struct LSurvCompResp{
E<:AbstractVector,
X<:AbstractVector,
Y<:AbstractVector,
W<:AbstractVector,
T<:Real,
I<:AbstractLSurvID,
V<:AbstractVector,
M<:AbstractMatrix,
} <: AbstractLSurvResp
enter::E
"`exit`: Time at observation end"
exit::X
"`y`: event type in observation (integer)"
y::Y
"`wts`: observation weights"
wts::W
"`eventtimes`: unique event times"
eventtimes::X
"`origin`: origin on the time scale"
origin::T
"`id`: person level identifier (must be wrapped in ID() function)"
id::Vector{I}
"`eventtypes`: vector of unique event types"
eventtypes::V
"`eventmatrix`: matrix of indicators on the observation level"
eventmatrix::M
end

LSurvCompResp(
enter::E,
exit::X,
y::Y,
wts::W,
id::Vector{I}
)

LSurvCompResp(
enter::E,
exit::X,
y::Y,
id::Vector{I}
)

LSurvCompResp(
enter::E,
exit::X,
y::Y,
wts::W,
)

LSurvCompResp(
enter::E,
exit::X,
y::Y,
)

LSurvival.LSurvResp — Type

Outcome type for survival outcome subject to left truncation and right censoring

struct LSurvResp{
E<:AbstractVector,
X<:AbstractVector,
Y<:AbstractVector,
W<:AbstractVector,
T<:Real,
I<:AbstractLSurvID,
} <: AbstractLSurvResp
enter::E
"`exit`: Time at observation end"
exit::X
"`y`: event occurrence in observation"
y::Y
"`wts`: observation weights"
wts::W
"`eventtimes`: unique event times"
eventtimes::E
"`origin`: origin on the time scale"
origin::T
"`id`: person level identifier (must be wrapped in ID() function)"
id::Vector{I}
end

function LSurvResp(
enter::E,
exit::X,
y::Y,
wts::W,
id::Vector{I},
) where {
E<:AbstractVector,
X<:AbstractVector,
Y<:AbstractVector,
W<:AbstractVector,
I<:AbstractLSurvID,
}

LSurvResp(
enter::E,
exit::X,
y::Y,
id::Vector{I},
) 

LSurvResp(
enter::E,
exit::X,
y::Y,
wts::W,
) 

LSurvResp(
enter::E,
exit::X,
y::Y,
)

LSurvResp(exit::X, y::Y) where {X<:AbstractVector,Y<:AbstractVector}

LSurvival.PHModel — Type

PHModel: Mutable object type for proportional hazards regression

mutable struct PHModel{G<:LSurvResp,L<:AbstractLSurvParms} <: AbstractPH
R::G        # Survival response
P::L        # parameters
ties::String #"efron" or"breslow"
fit::Bool
bh::AbstractMatrix
end

PHModel(
R::G,
P::L,
ties::String,
fit::Bool,
) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L, ties::String) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L) where {G<:LSurvResp,L<:AbstractLSurvParms}

Methods: fit, coef, confint, std_err, show

using LSurvival
using Random
import LSurvival: _stepcox!, dgm_comprisk

z,x,t,d, event,wt = dgm_comprisk(MersenneTwister(1212), 100);
enter = zeros(length(t));
X = hcat(x,z);
R = LSurvResp(enter, t, Int.(d), wt)
P = PHParms(X)
mf = PHModel(R,P)
 _fit!(mf)

LSurvival.PHModel — Method

PHModel: Mutable object type for proportional hazards regression

mutable struct PHModel{G<:LSurvResp,L<:AbstractLSurvParms} <: AbstractPH
R::G        # Survival response
P::L        # parameters
ties::String #"efron" or"breslow"
fit::Bool
bh::AbstractMatrix
end

PHModel(
R::G,
P::L,
ties::String,
fit::Bool,
) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L, ties::String) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L) where {G<:LSurvResp,L<:AbstractLSurvParms}

Methods: fit, coef, confint, std_err, show

using LSurvival
using Random
import LSurvival: _stepcox!, dgm_comprisk

z,x,t,d, event,wt = dgm_comprisk(MersenneTwister(1212), 100);
enter = zeros(length(t));
X = hcat(x,z);
R = LSurvResp(enter, t, Int.(d), wt)
P = PHParms(X)
mf = PHModel(R,P)
 _fit!(mf)

LSurvival.PHModel — Method

PHModel: Mutable object type for proportional hazards regression

mutable struct PHModel{G<:LSurvResp,L<:AbstractLSurvParms} <: AbstractPH
R::G        # Survival response
P::L        # parameters
ties::String #"efron" or"breslow"
fit::Bool
bh::AbstractMatrix
end

PHModel(
R::G,
P::L,
ties::String,
fit::Bool,
) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L, ties::String) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L) where {G<:LSurvResp,L<:AbstractLSurvParms}

Methods: fit, coef, confint, std_err, show

using LSurvival
using Random
import LSurvival: _stepcox!, dgm_comprisk

z,x,t,d, event,wt = dgm_comprisk(MersenneTwister(1212), 100);
enter = zeros(length(t));
X = hcat(x,z);
R = LSurvResp(enter, t, Int.(d), wt)
P = PHParms(X)
mf = PHModel(R,P)
 _fit!(mf)

LSurvival.PHModel — Method

PHModel: Mutable object type for proportional hazards regression

mutable struct PHModel{G<:LSurvResp,L<:AbstractLSurvParms} <: AbstractPH
R::G        # Survival response
P::L        # parameters
ties::String #"efron" or"breslow"
fit::Bool
bh::AbstractMatrix
end

PHModel(
R::G,
P::L,
ties::String,
fit::Bool,
) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L, ties::String) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L) where {G<:LSurvResp,L<:AbstractLSurvParms}

Methods: fit, coef, confint, std_err, show

using LSurvival
using Random
import LSurvival: _stepcox!, dgm_comprisk

z,x,t,d, event,wt = dgm_comprisk(MersenneTwister(1212), 100);
enter = zeros(length(t));
X = hcat(x,z);
R = LSurvResp(enter, t, Int.(d), wt)
P = PHParms(X)
mf = PHModel(R,P)
 _fit!(mf)

LSurvival.PHModel — Method

PHModel: Mutable object type for proportional hazards regression

mutable struct PHModel{G<:LSurvResp,L<:AbstractLSurvParms} <: AbstractPH
R::G        # Survival response
P::L        # parameters
ties::String #"efron" or"breslow"
fit::Bool
bh::AbstractMatrix
end

PHModel(
R::G,
P::L,
ties::String,
fit::Bool,
) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L, ties::String) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L) where {G<:LSurvResp,L<:AbstractLSurvParms}

Methods: fit, coef, confint, std_err, show

using LSurvival
using Random
import LSurvival: _stepcox!, dgm_comprisk

z,x,t,d, event,wt = dgm_comprisk(MersenneTwister(1212), 100);
enter = zeros(length(t));
X = hcat(x,z);
R = LSurvResp(enter, t, Int.(d), wt)
P = PHParms(X)
mf = PHModel(R,P)
 _fit!(mf)

LSurvival.PHModel — Method

PHModel: Mutable object type for proportional hazards regression

mutable struct PHModel{G<:LSurvResp,L<:AbstractLSurvParms} <: AbstractPH
R::G        # Survival response
P::L        # parameters
ties::String #"efron" or"breslow"
fit::Bool
bh::AbstractMatrix
end

PHModel(
R::G,
P::L,
ties::String,
fit::Bool,
) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L, ties::String) where {G<:LSurvResp,L<:AbstractLSurvParms}
PHModel(R::G, P::L) where {G<:LSurvResp,L<:AbstractLSurvParms}

Methods: fit, coef, confint, std_err, show

using LSurvival
using Random
import LSurvival: _stepcox!, dgm_comprisk

z,x,t,d, event,wt = dgm_comprisk(MersenneTwister(1212), 100);
enter = zeros(length(t));
X = hcat(x,z);
R = LSurvResp(enter, t, Int.(d), wt)
P = PHParms(X)
mf = PHModel(R,P)
 _fit!(mf)

LSurvival.PHSurv — Type

Mutable type for proportional hazards models

PHSsurv: Object type for proportional hazards regression

Methods: fit, show

mutable struct PHSurv{G<:Array{T} where {T<:PHModel}} <: AbstractNPSurv
fitlist::G        # Survival response
eventtypes::AbstractVector
times::AbstractVector
surv::Vector{Float64}
risk::Matrix{Float64}
basehaz::Vector{Float64}
event::Vector{Float64}
end

PHSurv(fitlist::Array{T}, eventtypes) where {T<:PHModel}
PHSurv(fitlist::Array{T}) where {T<:PHModel}

LSurvival.PHSurv — Method

Mutable type for proportional hazards models

PHSsurv: Object type for proportional hazards regression

Methods: fit, show

mutable struct PHSurv{G<:Array{T} where {T<:PHModel}} <: AbstractNPSurv
fitlist::G        # Survival response
eventtypes::AbstractVector
times::AbstractVector
surv::Vector{Float64}
risk::Matrix{Float64}
basehaz::Vector{Float64}
event::Vector{Float64}
end

PHSurv(fitlist::Array{T}, eventtypes) where {T<:PHModel}
PHSurv(fitlist::Array{T}) where {T<:PHModel}

LSurvival.PHSurv — Method

Mutable type for proportional hazards models

PHSsurv: Object type for proportional hazards regression

Methods: fit, show

mutable struct PHSurv{G<:Array{T} where {T<:PHModel}} <: AbstractNPSurv
fitlist::G        # Survival response
eventtypes::AbstractVector
times::AbstractVector
surv::Vector{Float64}
risk::Matrix{Float64}
basehaz::Vector{Float64}
event::Vector{Float64}
end

PHSurv(fitlist::Array{T}, eventtypes) where {T<:PHModel}
PHSurv(fitlist::Array{T}) where {T<:PHModel}

LSurvival.Strata — Type

Type for identifying individuals in survival outcomes.

Accepts any Number or String

[ID(i) for i in 1:10]

Used for the id argument in

Outcome types: LSurvResp, LSurvCompResp

Model types: PHModel, KMRisk, AJRisk

LSurvival._update_PHParms! — Method

calculate log likelihood, gradient, hessian at set value of coefficients

function _partial_LL!(
  m::M,
  # big indexes
  risksetidxs::Vector{Vector{T}},
  caseidxs::Vector{Vector{T}},
  ne::I,
  den::Vector{<:Real},
) where {M<:AbstractPH,I<:Int,T<:Int}

wrapper: calculate log partial likelihood, gradient, hessian contributions across all risk sets under a specified method for handling ties (efron and breslow estimators only)

p = size(X,2)
_LL = zeros(1)
_grad = zeros(p)
_hess = zeros(p,p)
_den = zeros(1)
#
_B = rand(p)
eventtimes = sort(unique(_out[findall(d.==1)]))

updatePHParms!(m, risksetidxs, caseidxs, ne, den)

LSurvival.aalen_johansen — Method

Aalen-Johansen estimator for cumulative risk

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

LSurvival.aj — Method

Deprecated function

Aalen-Johansen (survival) with late entry, possibly multiple observations per unit (simple function)

LSurvival.bootstrap — Method

Bootstrapping coefficients of a proportional hazards model

bootstrap(rng::MersenneTwister, m::PHModel)

using LSurvival, Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 500, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvResp(int, outt, d, ID.(id))    # specification with ID only
P = PHParms(X)

Mod = PHModel(R, P)
LSurvival._fit!(Mod, start=Mod.P._B)

# careful propogation of bootstrap sampling
idx, R2 = bootstrap(R)
P2 = bootstrap(idx, P)
Modb = PHModel(R2, P2)
LSurvival._fit!(Mod, start=Mod.P._B)

# convenience function for bootstrapping a model
Modc = bootstrap(Mod)
LSurvival._fit!(Modc, start=Modc.P._B)
Modc.P.X = nothing
Modc.R = nothing

bootstrap(rng::MersenneTwister, m::PHModel, iter::Int; kwargs...)

Bootstrap Cox model coefficients

LSurvival._fit!(mb, keepx=true, keepy=true, start=[0.0, 0.0])

using LSurvival, Random
res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
int = zeros(length(d)) # no late entry
X = hcat(z, x)

mainfit = fit(PHModel, X, int, outt, d .* (event .== 1), keepx=true, keepy=true)

mb = bootstrap(mainfit, 1000)
mainfit

LSurvival.bootstrap — Method

Bootstrapping coefficients of a proportional hazards model

bootstrap(rng::MersenneTwister, m::PHModel)

using LSurvival, Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 500, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvResp(int, outt, d, ID.(id))    # specification with ID only
P = PHParms(X)

Mod = PHModel(R, P)
LSurvival._fit!(Mod, start=Mod.P._B)

# careful propogation of bootstrap sampling
idx, R2 = bootstrap(R)
P2 = bootstrap(idx, P)
Modb = PHModel(R2, P2)
LSurvival._fit!(Mod, start=Mod.P._B)

# convenience function for bootstrapping a model
Modc = bootstrap(Mod)
LSurvival._fit!(Modc, start=Modc.P._B)
Modc.P.X = nothing
Modc.R = nothing

bootstrap(rng::MersenneTwister, m::PHModel, iter::Int; kwargs...)

Bootstrap Cox model coefficients

LSurvival._fit!(mb, keepx=true, keepy=true, start=[0.0, 0.0])

using LSurvival, Random
res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
int = zeros(length(d)) # no late entry
X = hcat(z, x)

mainfit = fit(PHModel, X, int, outt, d .* (event .== 1), keepx=true, keepy=true)

mb = bootstrap(mainfit, 1000)
mainfit

LSurvival.bootstrap — Method

using LSurvival, Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvResp(int, outt, d, ID.(id))    # specification with ID only
P = PHParms(X)
idx, R2 = bootstrap(R)
P2 = bootstrap(idx, P)

Mod = PHModel(R2, P2)
LSurvival._fit!(Mod, start=Mod.P._B)

LSurvival.bootstrap — Method

Bootstrap methods for Aalen-Johansen cumulative risk estimator

Returns:

If using bootstrap(fit): a single bootstrap draw

If using bootstrap(fit, 10) (e.g.): 10 bootstrap draws of the cumulative cause-specific risks at the end of follow up

using LSurvival
using Random

z, x, t, d, event, wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 100)
id = 1:length(x)
enter = zeros(length(t))

aj1 = aalen_johansen(enter, t, event, id=ID.(id), wts=wt)
aj2 = bootstrap(aj1, keepy=false);
ajboot = bootstrap(aj1, 10, keepy=false);
aj1

aj1.R
aj2.R

LSurvival.bootstrap — Method

Bootstrap methods for Kaplan-Meier survival curve estimator

Returns:

If using bootstrap(fit): a single bootstrap draw

If using bootstrap(fit, 10) (e.g.): 10 bootstrap draws of the survival probability at the end of follow up

using LSurvival
using Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
wts = rand(length(d))

km1 = kaplan_meier(int, outt, d, id=ID.(id), wts=wts)
km2 = bootstrap(km1, keepy=false)
km3 = bootstrap(km1, 10, keepy=false)
km1

km1.R
km2.R

LSurvival.bootstrap — Method

Bootstrap methods for Aalen-Johansen cumulative risk estimator

Returns:

If using bootstrap(fit): a single bootstrap draw

If using bootstrap(fit, 10) (e.g.): 10 bootstrap draws of the cumulative cause-specific risks at the end of follow up

using LSurvival
using Random

z, x, t, d, event, wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 100)
id = 1:length(x)
enter = zeros(length(t))

aj1 = aalen_johansen(enter, t, event, id=ID.(id), wts=wt)
aj2 = bootstrap(aj1, keepy=false);
ajboot = bootstrap(aj1, 10, keepy=false);
aj1

aj1.R
aj2.R

LSurvival.bootstrap — Method

Bootstrap methods for Kaplan-Meier survival curve estimator

Returns:

If using bootstrap(fit): a single bootstrap draw

If using bootstrap(fit, 10) (e.g.): 10 bootstrap draws of the survival probability at the end of follow up

using LSurvival
using Random

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
wts = rand(length(d))

km1 = kaplan_meier(int, outt, d, id=ID.(id), wts=wts)
km2 = bootstrap(km1, keepy=false)
km3 = bootstrap(km1, 10, keepy=false)
km1

km1.R
km2.R

LSurvival.bootstrap — Method

z,x,t,d,event,weights =
LSurvival.dgm_comprisk(MersenneTwister(1212), 300)
enter = zeros(length(event))

# survival outcome:
R = LSurvCompResp(enter, t, event, weights, ID.(collect(1:length(t))))    # specification with ID only
bootstrap(R) # note that entire observations/clusters identified by id are kept

LSurvival.bootstrap — Method

id, int, outt, data =
LSurvival.dgm(MersenneTwister(1212), 20, 5; afun = LSurvival.int_0)

d, X = data[:, 4], data[:, 1:3]
weights = rand(length(d))

# survival outcome:
R = LSurvResp(int, outt, d, ID.(id))    # specification with ID only

LSurvival.ci_from_coxmodels — Method

Deprecated function Estimating cumulative incidence from two or more cause-specific Cox models

z,x,outt,d,event,weights = LSurvival.dgm_comprisk(120) X = hcat(z,x) int = zeros(120) d1 = d .* Int.(event.== 1) d2 = d .* Int.(event.== 2) sum(d)/length(d)

lnhr1, ll1, g1, h1, bh1 = coxmodel(int, outt, d1, X, method="efron"); lnhr2, ll2, g2, h2, bh2 = coxmodel(int, outt, d2, X, method="efron"); bhlist = [bh1, bh2] coeflist = [lnhr1, lnhr2] covarmat = sum(X, dims=1) ./ size(X,1) ci, surv = cifromcoxmodels(bhlist;eventtypes=[1,2], coeflist=coeflist, covarmat=covarmat) ci, surv = cifromcoxmodels(bhlist;eventtypes=[1,2])

LSurvival.coxmodel — Method

Deprecated function

Estimate parameters of an extended Cox model

Using: Newton raphson algorithm with modified/adaptive step sizes

Keyword inputs: method="efron", inits=nothing , # initial parameter values, set to zero if this is nothing tol=10e-9, # convergence tolerance based on log likelihod: likrat = abs(lastLL/LL[1]), absdiff = abs(lastLL-LL[1]), reldiff = max(likrat, inv(likrat)) -1.0 maxiter=500 # maximum number of iterations for Newton Raphson algorithm (set to zero to calculate likelihood, gradient, Hessian at the initial parameter values)

Outputs: beta: coefficients ll: log partial likelihood history (all iterations), with final value being the (log) maximum partial likelihood (log-MPL) g: gradient vector (first derivative of log partial likelihood) at log-MPL h: hessian matrix (second derivative of log partial likelihood) at log-MPL basehaz: Matrix: baseline hazard at referent level of all covariates, weighted risk set size, weighted # of cases, time

Examples:

  using LSurvival
  # simulating discrete survival data for 20 individuals at 10 time points
  id, int, outt, data = LSurvival.dgm(20, 5;afun=LSurvival.int_0);
  
  d,X = data[:,4], data[:,1:3]
  
  # getting raw values of output
  args = (int, outt, d, X)
  beta, ll, g, h, basehaz = coxmodel(args..., method="efron")
  beta2, ll2, g2, h2, basehaz2 = coxmodel(args..., method="breslow")

  # nicer summary of results
  args = (int, outt, d, X)
  res = coxmodel(args..., method="efron");
  coxsum = cox_summary(res, alpha=0.05, verbose=true);
    

LSurvival.coxph — Method

fit for AbstractPH objects

 fit(::Type{M},
 X::AbstractMatrix,#{<:FP},
 enter::AbstractVector{<:Real},
 exit::AbstractVector{<:Real},
 y::Union{AbstractVector{<:Real},BitVector}
 ;
 ties ="breslow",
 wts::AbstractVector{<:Real}      = similar(y, 0),
 offset::AbstractVector{<:Real}   = similar(y, 0),
 fitargs...) where {M<:AbstractPH}
 ```

 ```{julia}
 using LSurvival
 using Random
 z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
 enter = zeros(length(t));
 X = hcat(x,rand(length(x)));
  #R = LSurvResp(enter, t, Int.(d), wt)
  #P = PHParms(X,"efron")
  #mod = PHModel(R,P, true)
  #_fit!(mod)
 m = fit(PHModel, X, enter, t, d, ties="efron")
 m2 = fit(PHModel, X, enter, t, d, ties="breslow")
 coeftable(m)
 ```

 ```{julia}
 using Random, LSurvival
    id, int, outt, dat =
        LSurvival.dgm(MersenneTwister(123123), 100, 100; afun = LSurvival.int_0)
    data = (
            int = int,
            outt = outt,
            d = dat[:,4] .== 1,
            x = dat[:,1],
            z = dat[:,2]
    )
       
    f = @formula(Surv(int, outt,d)~x+z)
    coxph(f, data)

LSurvival.coxph — Method

LSurvival.dgm — Method

Generating discrete survival data without competing risks

Usage: dgm(rng, n, maxT;afun=int0, yfun=yprob, lfun=lprob) dgm(n, maxT;afun=int0, yfun=yprob, lfun=lprob)

Where afun, yfun, and lfun are all functions that take arguments v,l,a and output time-specific values of a, y, and l respectively Example:

expit(mu) =  inv(1.0+exp(-mu))

function aprob(v,l,a)
expit(-1.0 + 3*v + 2*l)
end
  
function lprob(v,l,a)
expit(-3 + 2*v + 0*l + 0*a)
end
  
function yprob(v,l,a)
expit(-3 + 2*v + 0*l + 2*a)
end
  # 10 individuals followed for up to 5 times
LSurvival.dgm(10, 5;afun=aprob, yfun=yprob, lfun=lprob)

LSurvival.dgm_comprisk — Method

Generating continuous survival data with competing risks

Usage: dgmcomprisk(rng, n) dgmcomprisk(n)

    - rng = random number generator    
    - n = sample size

Example:

using LSurvival
    # 100 individuals with two competing events
z,x,t,d,event,weights = LSurvival.dgm_comprisk(100)
    

LSurvival.e_yearsoflifelost — Method

Deprecated function

Expected number of years of life lost due to cause k

using Distributions, Plots, Random
plotly()
z,x,t,d, event,weights = dgm_comprisk(n=200, rng=MersenneTwister(1232));
  
times_sd, cumhaz, ci_sd = subdistribution_hazard_cuminc(zeros(length(t)), t, event, dvalues=[1.0, 2.0]);
times_aj, S, ajest, riskset, events = aalen_johansen(zeros(length(t)), t, event, dvalues=[1.0, 2.0]);
time0, eyll0 = e_yearsoflifelost(times_aj, 1.0 .- S)  
time2, eyll1 = e_yearsoflifelost(times_aj, ajest[:,1])  
time1, eyll2 = e_yearsoflifelost(times_sd, ci_sd)  
  # CI estimates
plot(times_aj, ajest[:,1], label="AJ", st=:step);
plot!(times_sd, ci_sd, label="SD", st=:step)  
  # expected years of life lost by time k, given a specific cause or overall
plot(time0, eyll0, label="Overall", st=:step);
plot!(time1, eyll1, label="AJ", st=:step);
plot!(time2, eyll2, label="SD", st=:step) 

LSurvival.kaplan_meier — Method

fit for KMSurv objects

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)

or, equivalently:

kaplan_meier(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

LSurvival.km — Method

Deprecated function

Kaplan Meier with late entry, possibly multiple observations per unit (simple function)

LSurvival.km — Method

Deprecated function

Kaplan Meier for one observation per unit and no late entry (simple function)

LSurvival.lgh! — Method

deprecated function

LSurvival.lgh_breslow! — Method

Deprecated function

LSurvival.lgh_breslow! — Method

lgh_breslow!(_den, _LL, _grad, _hess, j, p, Xcases, Xriskset, _rcases, _rriskset, _wtcases, _wtriskset)

for a given risk set

#compute log-likelihood, gradient vector and hessian matrix of cox model given individual level contriubtions

Xcases=X[caseidx,:]
Xriskset=X[risksetidx,:]
 _rcases = _r[caseidx]
 _rriskset = _r[risksetidx]
 
 _wtcases=_wt[caseidx]
 _wtriskset=_wt[risksetidx]
p = size(X,2)
j = 1
 _LL = [0.0]
 _grad = zeros(p)
 _hess = zeros(p,p)
 _den = zeros(j)
lgh_breslow!(_den, _LL, _grad, _hess, j, p, Xcases, Xriskset, _rcases, _rriskset, _wtcases, _wtriskset)

LSurvival.lgh_efron! — Method

Deprecated function

LSurvival.lgh_efron! — Method

lgh_efron!(_den, _LL, _grad, _hess, j, p, Xcases, X, _rcases, _r, _wtcases, _wt, caseidx, risksetidx)

# for a given risk set
#compute log-likelihood, gradient vector and hessian matrix of cox model given individual level contriubtions
Xcases=X[caseidx,:]
Xriskset=X[risksetidx,:]
_rcases = _r[caseidx]
_rriskset = _r[risksetidx]

_wtcases=_wt[caseidx]
_wtriskset=_wt[risksetidx]
p = size(X,2)
j = 1
_LL = [0.0]
_grad = zeros(p)
_hess = zeros(p,p)
_den = zeros(j)
lgh_efron!(_den, _LL, _grad, _hess, j, p, Xcases, X, _rcases, _r, _wtcases, _wt, caseidx, risksetidx)

LSurvival.risk_from_coxphmodels — Method

fit for AJSurv objects

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

LSurvival.subdistribution_hazard_cuminc — Method

Deprecated function

Non-parametric sub-distribution hazard estimator estimating cumulative incidence via the subdistribution hazard function

Usage: subdistributionhazardcuminc(in,out,d;dvalues=[1.0, 2.0], weights=nothing, eps = 0.00000001)

in = time at entry (numeric vector)

out = time at exit (numeric vector)

d = event indicator (numeric or boolean vector)

keywords:

dvalues = (default = [1.0, 2.0]) a vector of the unique values of 'd' that indicate event types. By default, d is expected to take on values 0.0,1.0,2.0 for 3 event types (censored, event type 1, event type 2)

weights = vector of observation weights, or nothing (default)

eps = (default = 0.00000001) very small numeric value that helps in case of tied times that become misordered due to floating point errors

Output: tuple with entries

times: unique event times

cumhaz: cumulative subdistrution hazard for each event type

ci: Subdistrution hazard estimators of cumulative incidence for each event type. 1-sum of the CI for all event types is equal to overall survival.

events: number of events of each type used in calculating survival and cumulative incidence at each event time

names: vector of symbols [:times, :cumhaz, :ci] used as a mnemonic for the function output

Note: For time specific subdistribution hazard given by 'sdhaz(t)', the cumulative incidence for a specific event type calculated over time is

1.0 .- exp.(.-cumsum(sdhaz(t)))

Examples:

  using LSurvival, Random

  z,x,t,d, event,weights = LSurvival.dgm_comprisk(1000);
  
  # compare these two approaches, where Aalen-Johansen method requires having cause specific hazards for every event type
  times_sd, cumhaz, ci_sd = subdistribution_hazard_cuminc(zeros(length(t)), t, event, dvalues=[1.0, 2.0]);
  times_aj, surv, ajest, riskset, events = aalen_johansen(zeros(length(t)), t, event, dvalues=[1.0, 2.0]);
  

StatsAPI.confint — Method

Greenwood's formula for variance and confidence intervals of a Aalen-Johansen risk function

res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
int = zeros(length(d)) # no late entry
m = fit(AJSurv, int, outt, event)
stderror(m)
confint(m, level=0.95)

StatsAPI.confint — Method

Greenwood's formula for variance and confidence intervals of a Kaplan-Meier survival curve

StatsBase.stderror(m::KMSurv)

StatsBase.confint(m:KMSurv; level=0.95, method="normal")

method:

"normal" normality-based confidence intervals

"lognlog" log(-log(S(t))) based confidence intervals

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)
stderror(m)
confint(m, method="normal")
confint(m, method="lognlog") # log-log transformation

StatsAPI.fit! — Method

fit for KMSurv objects

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)

or, equivalently:

kaplan_meier(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

Aalen-Johansen estimator for cumulative risk

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

fit for AJSurv objects

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

StatsAPI.fit — Method

fit for AJSurv objects

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

StatsAPI.fit — Method

fit for AbstractPH objects

 fit(::Type{M},
 X::AbstractMatrix,#{<:FP},
 enter::AbstractVector{<:Real},
 exit::AbstractVector{<:Real},
 y::Union{AbstractVector{<:Real},BitVector}
 ;
 ties ="breslow",
 wts::AbstractVector{<:Real}      = similar(y, 0),
 offset::AbstractVector{<:Real}   = similar(y, 0),
 fitargs...) where {M<:AbstractPH}
 ```

 ```{julia}
 using LSurvival
 using Random
 z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
 enter = zeros(length(t));
 X = hcat(x,rand(length(x)));
  #R = LSurvResp(enter, t, Int.(d), wt)
  #P = PHParms(X,"efron")
  #mod = PHModel(R,P, true)
  #_fit!(mod)
 m = fit(PHModel, X, enter, t, d, ties="efron")
 m2 = fit(PHModel, X, enter, t, d, ties="breslow")
 coeftable(m)
 ```

 ```{julia}
 using Random, LSurvival
    id, int, outt, dat =
        LSurvival.dgm(MersenneTwister(123123), 100, 100; afun = LSurvival.int_0)
    data = (
            int = int,
            outt = outt,
            d = dat[:,4] .== 1,
            x = dat[:,1],
            z = dat[:,2]
    )
       
    f = @formula(Surv(int, outt,d)~x+z)
    coxph(f, data)

StatsAPI.fit — Method

Aalen-Johansen estimator for cumulative risk

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
   # event variable is coded 0[referent],1,2
m = fit(AJSurv, enter, t, event)
mw = fit(AJSurv, enter, t, event, wts=wt)

or, equivalently:

aalen_johansen(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

StatsAPI.fit — Method

fit for KMSurv objects

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)

or, equivalently:

kaplan_meier(enter::AbstractVector, exit::AbstractVector, y::AbstractVector,
   ; <keyword arguments>)

StatsAPI.stderror — Method

Greenwood's formula for variance and confidence intervals of a Aalen-Johansen risk function

res = z, x, outt, d, event, wts = LSurvival.dgm_comprisk(MersenneTwister(123123), 100)
int = zeros(length(d)) # no late entry
m = fit(AJSurv, int, outt, event)
stderror(m)
confint(m, level=0.95)

StatsAPI.stderror — Method

Greenwood's formula for variance and confidence intervals of a Kaplan-Meier survival curve

StatsBase.stderror(m::KMSurv)

StatsBase.confint(m:KMSurv; level=0.95, method="normal")

method:

"normal" normality-based confidence intervals

"lognlog" log(-log(S(t))) based confidence intervals

using LSurvival
using Random
z,x,t,d, event,wt = LSurvival.dgm_comprisk(MersenneTwister(1212), 1000);
enter = zeros(length(t));
m = fit(KMSurv, enter, t, d)
mw = fit(KMSurv, enter, t, d, wts=wt)
stderror(m)
confint(m, method="normal")
confint(m, method="lognlog") # log-log transformation